Move PermissionHandler into 'handlers' package
