+++ Project Walking/Source code/app/src/main/java/com/teamapricot/projectwalking/LocationHandler.java
@@ -6,6 +6,7 @@ import android.location.LocationManager;
 
 import androidx.appcompat.app.AppCompatActivity;
 
+import com.teamapricot.projectwalking.handlers.PermissionHandler;
 import com.teamapricot.projectwalking.view.dialogs.PermissionRejectedDialog;
 
 import java.util.ArrayList;
